backend/requirement.txt
fastapi
uvicorn
supabase
python-dotenv
openai
pydantic
google-generativeai
